Refactor styled, add datepicker and change container
