+++ src/screens/Recruiter/DateSelection.js
@@ -1,16 +1,13 @@
-import React from 'react';
-import {View} from 'react-native';
-import {
-  GeneralInput,
-  GeneralText,
-  GeneralPicker,
-  GeneralButton,
-} from '../../components/atoms';
+import React, {useState} from 'react';
+import {GeneralText, GeneralButton, Container} from '../../components/atoms';
 import {GeneralHeader} from '../../components/molecules';
+import DatePicker from 'react-native-date-picker';
+import {CenterView} from '../styled';
 
 export const DateSelection = () => {
+  const [date, setDate] = useState(new Date());
   return (
-    <View>
+    <Container>
       <GeneralHeader
         title="Contratar"
         isTabRendered={true}
@@ -18,13 +15,16 @@ export const DateSelection = () => {
         color="background"
         weight
       />
+      <CenterView>
         <GeneralText
           title="Seleccione fecha y horario"
           size="h1"
           color="secondary"
           weight
         />
+        <DatePicker mode="datetime" date={date} onDateChange={setDate} />
         <GeneralButton title="Siguiente" color="secondary" />
-    </View>
+      </CenterView>
+    </Container>
   );
 };
