+++ src/screens/styled.js
